requirements.txt
requests
gspread
google-auth
python-dotenv
pandas
boto3
finance-datareader
pandas
beautifulsoup4
plotly
numpy